Bug 537896: [HiDPI][GTK] SWT to receive/handle DPI change notification from OS
Connect the first Shell for a given Display to the "notify::scale-factor" signal,
in order to detect changes in DPI from GTK. Calculate whether the DPI has changed,
and notify listeners to the SWT.ZoomChanged event.

Tested on GTK3.24, Fedora 30, X11 and Wayland on Snippet373. No AllNonBrowser JUnit
tests fail.

Change-Id: I73623667386f447e2071c1beac46e3213fb2bda3
Signed-off-by: Eric Williams <ericwill@redhat.com>