Bug 550517: [GTK3] Swing components keep the focus when coming back to SWT widgets
Call proper constructor in XEmbeddedFrame which sets supportsXEmbed to true.
Coincidently the is the same solution to bug 548661. I have merged Nikita's
patch for bug 548661 (thanks Nikita!) and will upload only the snippet in
this patch.

Tested on GTK3.24 and Fedora 31 on X11 with the snippet attached. No
AllNonBrowser JUnit tests fail.

Change-Id: I8c2e30b90481c2b5bb108f447ad89b167a06f7d9
Signed-off-by: Eric Williams <ericwill@redhat.com>