Bug 522181: [GTK][WebKit2] Port clearSessions(), getCookie() and setCookie() to WebKit2
Implement set/get cookie on WebKit2. clearSessions() has already been ported so no work
is left to be done there.

The API introduced in WebKit2.20 allows for asynchronous setting/getting of cookies.
Because SWT expects a synchronous return, I've adapted the same approach to cookies
as Leo Ufimtsev did for getText() -- we spin the display loop and wait for the callback
to be triggered.

This patch also allows for test_get_set_Cookies to finally be enabled on WebKit2. Tested
on GTK3.24 and Fedora 31/30. No AllBrowserTests fail.

Change-Id: I1136e0c4de332b276b256bd7a57a1fb4ba019da7
Signed-off-by: Eric Williams <ericwill@redhat.com>