Bug 384851 - [CTabFolder] Top right control destroys border for last CTabItem in 4.2
If the setTopRight control with style SWT.FILL is the only control apart
from the actual CTabItems - i.e. no min/max button, no visible chevron -
the required spacing for the topRight control is missing.

The consequence is that the topRight control is moved left on top of the
last item or outside the tab folder bar if no item available. With an
item available this is barely visible because the spacing is only 3
pixel.

Change-Id: I51bd58c05b6ef0a8f89072d7de6e2ddd9ef64cea
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>