Bug 561444 - BadMatch X Error after Control.setRedraw(false)
Originally this fix intended to fix one crash, but eventually it turns
out to fix at least 4 connected bugs.

Two of them are when XIM input method is used. XIM doesn't really do
anything too wrong. What it does is calling `gdk_window_ensure_native()`,
which should be fine. Calling `gdk_window_ensure_native()` is enough to
get the same problems without XIM.

1) Drawing on a GC constructed on single-line Text crashes JVM.
2) Using single-line Text.setRedraw() crashes JVM.

Two others exist independent of XIM:

3) `Control.redrawWidget()` doesn't work for single-line Text.
4) After `Control.setRedraw(false)` + `Control.setRedraw(true)`,
   single-line Text stops reacting to mouse events.

(1)(2)(3) happen because SWT accesses private window of GtkEntry, and
in commit 596da843 (2010-10-21) GTK has changed this window to be a
`GDK_INPUT_ONLY` window (that is, non-paintable). As for (4) I'm not
sure, but since the same change fixes it too, I would expect it to
have the same root cause.

Trying to use `GC` on non-paintable window is an apparent wrongdoing.
However, SWT gets away with this while the non-paintable window
is NOT native. This is because non-native GDK windows do not exist on
Xorg side, and therefore inherit properties of their parent native
window. However, as soon as window becomes native, it also becomes
truly non-paintable, and trying to paint on it throws the `X Error`
that terminates JVM.

The key reason for accessing the private window in SWT was to get
`Text.setCursor()` to work. However, since Bug 124481 it is no longer
needed to do that in `paintWindow()`. Also, simply from the function
name it's already clear that it shouldn't return a non-paintable window.

Related history of SWT changes is:
Bug  22018 (2004-05-26) - `Text.paintWindow()` is changed make
  `Text.setCursor()` work. This is where SWT begins to access the
  private window.
Bug 124481 (2006-02-01) - `eventWindow()` is implemented
  Due to oversight, `Text.setCursor()` no longer uses custom
  `Text.paintWindow()`.
Bug 146575 (2006-06-15) - Again fixes `Text.setCursor()`
  `Text.eventWindow()` is implemented to once again return
  `Text.paintWindow()`.
Bug 281352 (2009-07-17) - `Text.paintWindow()` is changed to support
  `SWT.SEARCH`. `Text.paintWindow()` is modified to find the last child
   instead of the first one.

This change also removes GTK4-specific `Text.paintSurface()`, which is
probably wrong. I didn't test that myself.

Change-Id: I8ad1077ef2b6e2f54460912fdce1905364346dd6
Signed-off-by: Alexandr Miloslavskiy <alexandr.miloslavskiy@syntevo.com>