Fix for bug 562126 SWT Toolbar not Visible with Certain Scaling Factors (HiDPI)
The option TBSTYLE_EX_HIDECLIPPEDBUTTONS hides toolbar items completely instead of clipping them.
A slight reduction in size can happen on certain scaling factors for a sequence of computeSize, setSize on certain scaling factors as
described in the bug description. I think it is in general better to show a slightly clipped button than nothing at all, but this
should be reviewed.

Change-Id: If70bac6b8efd8a33f58fa6b5468d9466fe0d0ad1
Signed-off-by: Thomas Rohwer <trohwer85@gmail.com>