Bug 499215 - Chevron is wrapped to new line
The failure can also appear when min or max button is added to a visible
CTabFolder.
Precondition for this failure is that all CTabItems are smaller in
height than the icon + margins. When the min/max is to be shown the
following happened before this change:
 - tab bar height is recalculated
 - min/max button is created and bounds calculated (at once)

The fix for that scenario is to split the one button handling method
into a button creation and bounds calculation method and reorder the
invocation to: create button, updateTabHeight (which now consider the
created buttons) and set button bounds.

Situation for chevron is a bit more complicate because the decision if a
chevron is shown depends on the item size calculation which already
require the tab bar height.
The fix is similar to the previous. If chevron is added/removed the tab
bar height has to be recalculated for the case it changed from the
changed chevron.

Change-Id: I1fffac97333c4414ca0dc1394ea87377f12e74fb
Signed-off-by: Paul Pazderski <paul-eclipse@ppazderski.de>