Bug 563555 - [GTK] Closing Shell shows annoying "is ready" popup on GNOME
To my understanding, the problem here is that `gdk_window_focus()` gets called
before the closed Shell is disposed. Moving `fixActiveShell()` below
`super.dispose()` works fine (if `fixActiveShell()` is fixed to avoid NPE).

However, I found out that entire `fixActiveShell()` can simply be removed.
It was added in Bug 135946 as workaround for GTK bug, which is said to be fixed
in GTK 2.6.8. Since we're only supporting GTK 3.20 at the moment, it is no
longer needed.

I have tested using the snippet from this patch and for me, GTK always
activates the parent Shell of closed/hidden Shell, that is it does the same
that `fixActiveShell()` tried to do.

Change-Id: I2ffc7bf81cf4c6a7b117a819e4006fe1fef63807
Signed-off-by: Alexandr Miloslavskiy <alexandr.miloslavskiy@syntevo.com>