Bug 564097 - [GTK][HiDPI] Cairo auto scaling causing scaling problems
Removed unused scaling factor variable & gsettingsProc in Device.
Applied dpi_changed signal to all shells created, allowing Eclipse to
detect when zoom has changed & relayout. Before the signal was connected
only to the first shell of a Display. (I'm guessing that Eclipse runs
more than one shell and the first shell actually never changes monitors
-> never changes DPI when user moves the Workbench to a higher DPI
monitor)

On DpiChanged signal emitted, deviceZoom is actually changed to the
correct value rather than the scale factor of the point (0, 0).

On first shell creation, get the scale factor on the monitor the shell
is created on. This ensures that the icons/images are scaled properly.

Change-Id: I9a554cf605ce8e5d96fdcc31f355a242b5511975
Signed-off-by: Paul D'Pong <sdamrong@redhat.com>