Bug 564504 - Fix scaling
Not sure how it is supposed to be. In
https://git.eclipse.org/r/plugins/gitiles/platform/eclipse.platform.swt/+/e02d49aefe42ac4c77b81048299ab069ddb5c2ba%5E%21/#F33
definitely the changes on getBounds() and getCurrentPoint() do not apply
the scaling to the float array. Maybe the methods are never called and
can be removed completely? Or you *must* not apply it? Maybe they can be
remove but they are public so api...

Change-Id: I830fa135ad1a20ffc975e60033e3b39261dd879a
Signed-off-by: Carsten Hammer <carsten.hammer@t-online.de>