Bug 565613 - [Win32] Scrollbar thumb "jumps" when dragging it
The problem was that `RDW_FRAME` caused scrollbars to redraw. When
scrollbar's (max-min) is not very high, there's an easily visible
distance between two distinct scrollbar positions. When scrollbar is
forced to paint, it paints a distinct position instead of current
dragging position, causing a visible "jump" of thumb.

`RDW_FRAME` has been there since first SWT commit. I doubt that it's
ever necessary in SWT. Even if it is, using it for every redraw is
definitely an overkill. On top of causing visual glitches such as this
Bug, it is also a performance issue, causing unnecessary drawing on
every update. This is even more so because SWT handles `WM_NCPAINT` to
overpaint some borders.

This patch has a potential of raising some new minor painting problems.
Specifically, it's possible that in some circumstances, windows's border
will not be updated correctly. I don't expect that such problems will
arise. However, if they do, it should be easy to address them surgically
instead of using `RDW_FRAME` for every update.

Change-Id: Iea6f8db8b98ae0e1f3ef460d7285de8bee215d0b
Signed-off-by: Alexandr Miloslavskiy <alexandr.miloslavskiy@syntevo.com>