Bug 567826 - [GTK] JVM crash on Display.dispose() due to Clipboard
If 'SWT.Dispose' handler is a bit slow, user might be able to switch to
other application and Ctrl+C there. When this happens, Linux calls
'ClipboardProxy.clearFunc()' with the following stack:

  Native: clipboard_unset() at gtk/gtkclipboard.c:782
  Native: selection_clear_event_cb() at gtk/gtkclipboard.c:424
  Native: <emit signal for 'GDK_SELECTION_CLEAR'>
  Native: gtk_widget_event_internal() at gtk/gtkwidget.c:7808
  Native: gtk_main_do_event() at gtk/gtkmain.c:1860
  Java:   org.eclipse.swt.internal.gtk.GTK._gtk_main_do_event()
  Java:   org.eclipse.swt.widgets.Display.eventProc()
  Java:   org.eclipse.swt.internal.gtk.OS._g_main_context_iteration()
  Java:   org.eclipse.swt.widgets.Display.readAndDispatch()
  Java:   org.eclipse.swt.widgets.Display.release()

If 'Callback' for 'clearFunc' is already disposed, JVM crashes.

The solution is to dispose callback after 'Display.readAndDispatch()' in
'Display.release()'.

This patch has a curious side effect: if 'SWT.Dispose' handler is slow,
Shell is now visible (and eventually shows "not responding") for its
duration. I understand that this is because 'gtk_clipboard_store()'
also does some unexpected event processing. I consider this side effect
as not a regression because the same behavior is seen when clipboard is
not owned by SWT by the time of 'Display.dispose()'.

Change-Id: I2cc11600f06030243e55d70f24b881dcb9b67503
Signed-off-by: Alexandr Miloslavskiy <alexandr.miloslavskiy@syntevo.com>