Bug 565969: [Big Sur] Table selection drawn over text/checkbox when SWT.EraseItem is hooked
In Table/Tree, highlightSelectionInClipRect() call from Cocoa is ignored when SWT.EraseItem is hooked and selection is drawn in drawInteriorWithFrame_inView.
On BigSur, calling highlightSelectionInClipRect in drawInteriorWithFrame_inView draws over the full row and not just the cellRect.
This causes drawing over other cells content. Workaround is to draw the highlight background ourselves and not call highlightSelectionInClipRect to draw it.

Change-Id: I9f2e5e93730a0d63221eb628807d0fd7567deba7