Bug 569147 - [BigSur] Control#update() causes losing dirty state without displaying proper visual content with specific conditions.
Bigsur seems to force the use of the Automatic Deferred Painting (ADP)
mechanism, so the previous fix modified the code in update() to do
nothing on BigSur. This behaviour is however not expected on apps linked
with SDK < 10.14 and could be 'fixed' or changed.

Adding an internal system property, just in case that happens.

Change-Id: Iced9c2aca2e9fb7ca37a9e213cf8d8983d5bdf59