Bug 569529 - [Cocoa] Link and multi-line Text hog CPU in dark theme
The key change is to use 'lastAppAppearance' to avoid endless redraws.

Tests for 'APPEARANCE.Dark' were removed in anticipation of fixing
'OS.setTheme()' to be able to switch theme dynamically. I understand
that colors need to be updated again when switching fro mDark to Light.

Colors are updated *before* calling 'super', because it's only
reasonable to do the drawing with new colors already.

'Text.drawBackground()' is called from 'Text.drawRect()', so it doesn't
need a color adjustment of its own.

Change-Id: If504c4dd5da1a0984f75829e81efa507d41763d4
Signed-off-by: Alexandr Miloslavskiy <alexandr.miloslavskiy@syntevo.com>