Bug 552063 - [Cocoa] Menu not clickable on macOS at launch
When Eclipse launches, the menu is not immediately clickable. A recent
change in macOS broke the functionality used before, which calls
`activateWithOptions` outside of the `applicationDidFinishLaunching`
delegate call.

As per https://stackoverflow.com/questions/62739862 this can be fixed by
moving the `activateWithOptions` call into a new handler for the
delegate.

On macOS 10.15 (but not 11.1), `setActivationPolicy` also has to be
moved to `applicationDidFinishLaunching`.

The erroneous behaviour is still seen when the splash screen is shown,
but when running Eclipse with `-nosplash` the menu is clickable with
this fix.

Change-Id: Ibfffab20b66e843cc507a1982dc16f0f257cf92a
Signed-off-by: Alex Blewitt <alex.blewitt@gmail.com>