Bug 570178 - [GTK] Fix cast type for gtk_im_context_filter_keypress()
The problem here is that GTK3 wants 'GdkEventKey*' while GTK4 doesn't
know this type at all and wants 'GdkEvent*'. So whatever is chosen, the
type will be wrong on one of GTK versions. The type itself is not
important for the compiled binary, it's merely a compilation issue.

At the same time, final patch for Bug 570178 turned all warnings into
errors to make sure that significant warnings are not overlooked.

Until Bug 570533 is resolved, a temporary hack is applied to allow GTK4
to build.

Change-Id: Ica2dd032ad5c6c60ec575a3e7b7db97b9cdf1119
Signed-off-by: Alexander Kurtakov <akurtako@redhat.com>
Signed-off-by: Alexandr Miloslavskiy <alexandr.miloslavskiy@syntevo.com>