Bug 569468 - [GTK3] SWT/GTK3 crash when using GTKLookAndFeel in Swing
Swing GTKLookAndFeel relies on the deprecated GDK lock to synchronize
GTK function calls from the Event Dispatch Thread.

To avoid data races, memory corruption and chrashes when using
GTKLookAndFeel, SWT also has to synchronize with the GDK lock.
Use GTK event loop functions that perform the necessary locking and
unlocking automatically.

To allow Swing objects to be created on the main thread, GTK lock has
to be reentrant. Since the default lock isn't, replace it with GRecMutex.

With GDK locking in place, EventQueue.invokeAndWait called from
the main thread is very likely to deadlock and should be avoided.

Change-Id: I3b217037f515cb695c1c3ef63387d218b559a617
Signed-off-by: Nikita Nemkin <nikita@nemkin.ru>