Bug 570502 - [GTK3] Readonly Combo ignored set background color
The problem in code before patch is: if 'setForeground()' is called
first, it eventually leads to 'gtk_css_provider_load_from_css()' being
called with context for 'handle' instead of 'buttonHandle'. When
'setBackground()' is called later, it ignores correct 'context' because
'provider' is already allocated.

The patch reworks code to a simpler approach: a single function
'updateCss()' which handles generating css for all cases. All other
code paths simply invoke this central function. For further simplicity,
the same css provider is used everywhere.

Benefit#1 is that there is no longer any potential to use the wrong
context/provider, such as in the bug fixed here.

Benefit#2 is that '.setBackgroundGdkRGBA()' no longer needs to guess
the old color when called with rgba=null. Instead, it can simply omit
this part in the new css, which will cause GTK to use default color
again.

Benefit#3 is that it is no longer needed to cache 'cssButtonBackground'
and 'cssButtonForeground'. By reducing the state, code becomes easier
to read and more reliable.

This patch also does trivial refactoring in 'releaseHandle()'. As a
result, it becomes obvious that 'entryHandle' was not released in old
code. Not sure if it's a bug, and I really want to finish this patch
already, so it's left for someone else to figure.

Old code had 'setBackgroundGdkRGBA (fixedHandle, rgba)'. I don't think
that it's useful in any way, so I removed it.

Change-Id: I3dc79a4730d09706385c18b8c25ec2b79329ffc4
Signed-off-by: Alexandr Miloslavskiy <alexandr.miloslavskiy@syntevo.com>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.swt/+/176058
Tested-by: Platform Bot <platform-bot@eclipse.org>
Reviewed-by: Alexander Kurtakov <akurtako@redhat.com>