Bug 570533 - [GTK4] Split gtk_im_context_filter_keypress
This patch splits implementations for gtk_im_context_filter_keypress().

The problem with it is that in GTK4, the type of the argument was
changed from 'GdkEventKey*' to 'GdkEvent*'.

Since Bug 570178, all C warnings are treated as errors, which requires
a precise argument type. Solve this problem by making two
implementations.

Change-Id: I6d8c32df273b90d87bc65d6e9ac610a29eb2a431
Signed-off-by: Alexandr Miloslavskiy <alexandr.miloslavskiy@syntevo.com>