Bug 567754 - File dialog does not change file name when switching file type
As of macOS 10.15, all apps have the sandbox versions of the Open/Save
panels even if they are not sandboxed. This means you can't
setNameFieldStringValue; an unsecure app might change the filename when
the user was not expecting it.

Use setAllowedFileType for NSSavePanel, instead of trying to explicitly
append the selected extension.

The side effect of using setAllowedFileType is that if a multi-part
extension (such as tar.gz) is selected in Save Dialog's extension
pop-up, only the last part (gz) is considered as the extension. This is
because, Mac doesn't support multi-part extensions, it only
considers the last part of the extension.

Change-Id: Ife16e33b8f0d75b384d46c2db8d2effd4ffd43b9