Bug 568185 - [GTK4] Port SWT Tree snippets - getWidth
- Removed useFixedWidth boolean
- Removal is to allow GTK4 to have the same behavior as GTK3 for the
getWidth function
- This believe this change is good and okay in GTK3 because getWidth
should be returning the display width of the column rather than the
"fixed-width" which may not have changed due to size allocation that is
not controlled by the GtkTreeColumn. The "width" field is just more
flexible and does not require a field that is changed in the
size-allocate handler.

Change-Id: I66e5d49dcea96f975e0658081699ecda337adc7b
Signed-off-by: Paul D'Pong <sdamrong@redhat.com>