Bug 571746 - Prevent execution of JavaScript on close when disabled
There is a race condition because jsEnabled is updated asynchronously.
This might result in a situation when it's still true and the browser is
requested to be closed although it was never fully visible. This results
in a deadlock with Content Assist. See bug 567881 for a stacktrace
exposing the issue.

Change-Id: I13e56ff056dbf84b9a7cd7afb37bdcc42c9b88ce
Signed-off-by: Gunnar Wagenknecht <gunnar@wagenknecht.org>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.swt/+/177319
Reviewed-by: Sravan Kumar Lakkimsetti <sravankumarl@in.ibm.com>
Tested-by: Sravan Kumar Lakkimsetti <sravankumarl@in.ibm.com>