Bug 571033 - [GTK] Clean up ToolTips Implementation
- Moved setToolTipText utility function from Shell to Widget, as there
is no dependency on Shell required and the function is mainly used in
more broad cases. This results in cleaner code as there is no need to
get the Shell object.

- Remove fixChildren overrides in Table, Tree, as the changes in the
Shell or underlying GtkWindow does not affect the call to
gtk_widget_set_tooltip_text. This seems like a legacy thing that SWT
used to do when GtkWidget did not have native tooltips.

Change-Id: I8b886cbc34e3c5b0e1bb18b30a3a1f4edbacde24
Signed-off-by: Paul D'Pong <sdamrong@redhat.com>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.swt/+/181797
Tested-by: Platform Bot <platform-bot@eclipse.org>
Reviewed-by: Alexander Kurtakov <akurtako@redhat.com>