Bug 545846 - StyledText: wrong word selected on double-click
This resurrects Thomas Singer's commit cafa676acdbc, which had been
reverted in commit 29806b6c3cab.

Amend that code to not trigger if the mouse click occurs beyond the
end of the line or on a whitespace character, so in those cases, the
behavior is unchanged. Also keep the selection code the same so that
it always first finds the start, then the end of the word.

Also add tests for word selection on double click.

This resolves the problem that led to the revert: {} block selections
on double click didn't work anymore in Java and other editors. With
this modification, they do. Word selection on double click in a pure
SWT StyledText works, and double-click behavior in JFace-enhanced
StyledTexts in SourceViewers with ICharacterPairMatchers is unchanged
as far as I see.

Change-Id: I38258c001576ebb2c43caa95d0e0dc32143a2928
Also-by: Thomas Singer <ts-swt@syntevo.com>
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.swt/+/183868
Tested-by: Platform Bot <platform-bot@eclipse.org>
Reviewed-by: Mickael Istria <mistria@redhat.com>