Bug 501491 - [Graphics] Replace custom drawing on close button
Eclipse's close button in non-hot mode looks should look close enough to
whatever is used normally on the corresponding platform to not stick
out.

The current close button of CTabFolder fails by this measure, on all
three platforms.

Native look on Mac for closeable tabs appears to be a plain cross, as
can be seen in the Finder's tabs. As hover effect, the background is
made (in light mode) a darker gray. (Darker gray rectangle with rounded
corners).

Native look on Windows is also a plain cross, with the same hover
effect. See for instance tabs in Microsoft Edge. Other web browsers seem
to have converged on this same UI across platforms. On Windows 10, the
shell close button is also such a thin plain cross, as hover effect, the
button gets a glaringly red background. The tab close button's cross is
smaller, but appears to use also in non-hot mode a slightly thicker
lines.

Native look on GTK/Adwaita on CentOS 7 for tabs appears to be also a
plain cross, but with thicker lines, as can be seen for instance in
gedit with multiple files open. As hover effect, the background becomes
a brighter shade of gray. The shell close icon uses the same cross.
Firefox uses the same cross and hover effect (darker gray) as on other
platforms.

This change updates the drawing of the close icon to look similar to the
platform icons. It also moves closer to popular tools like the Chrome
browser or development tools like VScode.


Change-Id: Iddbb709c5f291a1e7a52b3e77c46da605d4ad8a8
Signed-off-by: Lars Vogel <Lars.Vogel@vogella.com>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.swt/+/183887
Tested-by: Platform Bot <platform-bot@eclipse.org>