Bug 575787 - [GTK] Drop fixAccessibility() hack in Table, Tree
In GTK2, there was a problem where inserting many items in Table/Tree
was super slow if Screen Reader or other accessibility was active.

A lot of time was spent in this stack:
  count_rows
  get_row_count
  gail_tree_view_get_n_children
  gail_tree_view_ref_child
  spi_atk_bridge_signal_listener
  <...>
  model_row_inserted
  <...>
  _gtk_tree_model_row_inserted

Basically, on every row inserted, it counted all other rows in Tree,
making it an O(N*N), probably with large constant.

However, in GTK 3.3.4 this was abolished in commit:
  0c2f454b by Benjamin Otte, 2011-07-07 11:54:04
  a11y: Re-implement gtk_tree_view_accessible_get_n_rows()

Nowadays, the hack makes things slower, not faster. Let's get rid of it.

Performance on my machine for 100_000 items:
[S] = Screen reader active [P] = With patch
   [--] 1.70sec=Tree(REGULAR) 1.82sec=Tree(VIRTUAL) 0.84sec=Table(REGULAR) 1.02sec=Table(VIRTUAL)
   [-P] 1.40sec=Tree(REGULAR) 1.27sec=Tree(VIRTUAL) 0.59sec=Table(REGULAR) 0.55sec=Table(VIRTUAL)
   [S-] 3.43sec=Tree(REGULAR) 3.49sec=Tree(VIRTUAL) 0.82sec=Table(REGULAR) 0.99sec=Table(VIRTUAL)
   [SP] 3.13sec=Tree(REGULAR) 2.94sec=Tree(VIRTUAL) 0.59sec=Table(REGULAR) 0.50sec=Table(VIRTUAL)

I have tested it with Orca screen reader (preinstalled in Ubuntu).

Signed-off-by: Alexandr Miloslavskiy <alexandr.miloslavskiy@syntevo.com>
Change-Id: Ic711facd75b62498030d4ee967a2774bb2d71713
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.swt/+/185276
Tested-by: Alexander Kurtakov <akurtako@redhat.com>
Tested-by: Platform Bot <platform-bot@eclipse.org>
Reviewed-by: Alexander Kurtakov <akurtako@redhat.com>