Bug 576215 - [Wayland] Mouse events not received as on other platforms
This patch addresses mouse events not being the same between x11 and
Wayland. The issue was MouseMove events were being left out of the queue
which prevented them from being sent out. Passes AllNonBrowserTests and
DnD example works as expected.

The Wayland behavior is not 100% the same as x11 due to the initial
MouseDown event being cached before being sent out. However, as soon as
the mouse begins to move, the down message pops up, and the behavior is
as expected.

Change-Id: I090953be3bf592c4161ec497a53e47964904fe45
Signed-off-by: jmajano <jmajano@redhat.com>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.swt/+/186019
Tested-by: Platform Bot <platform-bot@eclipse.org>
Reviewed-by: Alexander Kurtakov <akurtako@redhat.com>