Bug 576655 - Use ResourceTracker in JUnit tests
The goal of this change is to automatically detect leaks in SWT
internal code such as leak in CTabFolder.setFont() that was fixed in
569752. Such errors happen rarely, but on the other hand, detecting
them is cheap, so it makes sense to detect.

I decided to enable tracking in 'AllNonBrowserTests' because it's often
chosen over 'AllTests'.

Also fixes some existing leaks in JUnit tests.

Change-Id: Ida9c2a0b31465546adedbe6bd1850d1d9ac618e7
Signed-off-by: Alexandr Miloslavskiy <alexandr.miloslavskiy@syntevo.com>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.swt/+/186543
Tested-by: Platform Bot <platform-bot@eclipse.org>
Reviewed-by: Andrey Loskutov <loskutov@gmx.de>