Bug 575787 - Fix incorrect use of TVGN_ROOT
'TVGN_ROOT' is a constant to be used together with 'TVM_GETNEXTITEM'.
The correct constant to use in place of 'HTREEITEM' is 'TVI_ROOT'.
And these two have different values!

Coincidentally though, 'TVGN_ROOT' is 0, which in case of 'HTREEITEM' is
seen as NULL, which in many APIs can be used instead of 'TVI_ROOT' to
mention root item. That's why things kind of worked.

However, some APIs can't handle root items. For example,
'TVM_GETITEMSTATE'. In case of 0, it silently fails and returns 0,
that's why it went unnoticed. However, in case of 'TVI_ROOT' it crashes,
that's why guards are added in this patch to make sure it's not called.

This patch fixes code to avoid confusion for the readers. I myself was
quite surprised to see 'TVGN_ROOT' where it shouldn't be.

The patch also simplifies 'Tree.setItemCount()'. Previously,
(OS.TVM_GETNEXTITEM, OS.TVGN_ROOT) was used to get the first child of
the Tree, which was rather confusing. 'OS.TVGN_CHILD' is more clear.
Windows handles both combinations in the same way in case of root item.

Change-Id: Ieca8e00d183b792a43b07b4be740580542763992
Signed-off-by: Alexandr Miloslavskiy <alexandr.miloslavskiy@syntevo.com>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.swt/+/187202
Tested-by: Niraj Modi <niraj.modi@in.ibm.com>
Reviewed-by: Niraj Modi <niraj.modi@in.ibm.com>