Bug 577431 - TabFolder not taking full width in GTK4
The first part of the solution was to connect to the "notify::maximized"
signal, which would then allow the code responsible for resizing while
using the "FillLayout" layout to run.

Next, in GTK4 using gtk_window_get_default_size returns the previously
set size even if the window is maximized. As a result, while using
"FillLayout", the widgets were not being resized correctly. After
connecting the maximized signal, SWT now knows that it is maximized, so
all we have to do is fetch the monitor geometry, and use that to
calculate the dimensions ONLY while maximized.

Read comment in code for further details.

Change-Id: I5f809af60bd3c02583214de7183aaaaf26eda1cb
Signed-off-by: Joel Majano <jmajano@redhat.com>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.swt/+/188297
Tested-by: Platform Bot <platform-bot@eclipse.org>
Tested-by: Alexander Kurtakov <akurtako@redhat.com>
Reviewed-by: Alexander Kurtakov <akurtako@redhat.com>