Bug 567215 - SWT Combo drop down background not set correctly
menuHandle and popupHandle were not being set in GTK4 which meant that
when updateCSS() was trying to set the CSS for menuHandle, nothing was
happening.The fix is to map menuHandle to the correct widget which can
be done based on the CSS hierachy. Simply find the GtkPopover which is
the parent of the drop down menu items then get its child, "contents"
which is the parent that needs its background set.

One outstanding issue is that popupHandle is not mapped. This is not
straightforward to fix due to popupHandle being used for more than just
setting CSS properties. It is used to add items and other miscellaneous
things, which means it will have be to revisited later.

Change-Id: I3eef34623bca21e9074ec7be49830e222c01c7be
Signed-off-by: Joel Majano <jmajano@redhat.com>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.swt/+/188489
Tested-by: Platform Bot <platform-bot@eclipse.org>
Reviewed-by: Alexander Kurtakov <akurtako@redhat.com>