Bug 577858 - [cocoa] Table/Tree drag&drop incorrectly starts even if DND.DragStart returns event.doit=false
The side effect of this patch is that when drag is canceled, moving
mouse while holding the button selects items under it. This is
different from what happens on Windows/Linux, where canceled drag
ignores further mouse movements. But oh well, I guess that's what
macOS chooses to do in this case.

Apple documentation for 'outlineView:writeItems:toPasteboard:' says [1]:
  To refuse the drag, return NO.
  To start a drag, return YES and place the drag data onto the pboard

[1] https://developer.apple.com/documentation/appkit/nsoutlineviewdatasource/1532910-outlineview?language=objc

Change-Id: I2c1e997ba5e6195b9758c06ea9bf0496ef7c7a56
Signed-off-by: Alexandr Miloslavskiy <alexandr.miloslavskiy@syntevo.com>
Reviewed-on: https://git.eclipse.org/r/c/platform/eclipse.platform.swt/+/188976
Tested-by: Platform Bot <platform-bot@eclipse.org>
Tested-by: Lakshmi P Shanmugam <lshanmug@in.ibm.com>
Reviewed-by: Lakshmi P Shanmugam <lshanmug@in.ibm.com>